pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-registry"
version = "0.1.0"
description = "A comprehensive prompt management system with MCP server integration"
authors = [
    { name = "Prompt Registry Team", email = "team@promptregistry.dev" }
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.0",
    "semantic-version>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
prompt-registry = "prompt_registry.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["prompt_registry"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
